add keep_sequence to re-use parquet reader

+++ src/metric_engine/src/read.rs
@@ -104,6 +104,8 @@ pub(crate) struct MergeExec {
     seq_idx: usize,
     /// Operator to merge values when primary keys are the same
     value_operator: Arc<dyn MergeOperator>,
+    /// Whether to keep the sequence column in the output
+    keep_sequence: bool,
 }
 
 impl MergeExec {
@@ -112,12 +114,14 @@ impl MergeExec {
         num_primary_keys: usize,
         seq_idx: usize,
         value_operator: Arc<dyn MergeOperator>,
+        keep_sequence: bool,
     ) -> Self {
         Self {
             input,
             num_primary_keys,
             seq_idx,
             value_operator,
+            keep_sequence,
         }
     }
 }
@@ -129,8 +133,8 @@ impl DisplayAs for MergeExec {
     ) -> std::fmt::Result {
         write!(
             f,
-            "MergeExec: [primary_keys: {}, seq_idx: {}]",
-            self.num_primary_keys, self.seq_idx
+            "MergeExec: [primary_keys: {}, seq_idx: {}, keep_sequence: {}]",
+            self.num_primary_keys, self.seq_idx, self.keep_sequence
         )?;
         Ok(())
     }
@@ -170,6 +174,7 @@ impl ExecutionPlan for MergeExec {
             self.num_primary_keys,
             self.seq_idx,
             self.value_operator.clone(),
+            self.keep_sequence,
         )))
     }
 
@@ -187,6 +192,7 @@ impl ExecutionPlan for MergeExec {
             self.num_primary_keys,
             self.seq_idx,
             self.value_operator.clone(),
+            self.keep_sequence,
         )))
     }
 }
@@ -196,6 +202,7 @@ struct MergeStream {
     num_primary_keys: usize,
     seq_idx: usize,
     value_operator: MergeOperatorRef,
+    keep_sequence: bool,
 
     pending_batch: Option<RecordBatch>,
     arrow_schema: SchemaRef,
@@ -207,28 +214,34 @@ impl MergeStream {
         num_primary_keys: usize,
         seq_idx: usize,
         value_operator: MergeOperatorRef,
+        keep_sequence: bool,
     ) -> Self {
-        let fields = stream
-            .schema()
-            .fields()
-            .into_iter()
-            .filter_map(|f| {
-                if f.name() == SEQ_COLUMN_NAME {
-                    None
-                } else {
-                    Some(f.clone())
-                }
-            })
-            .collect_vec();
-        let arrow_schema = Arc::new(Schema::new_with_metadata(
-            fields,
-            stream.schema().metadata.clone(),
-        ));
+        let arrow_schema = if keep_sequence {
+            let fields = stream
+                .schema()
+                .fields()
+                .into_iter()
+                .filter_map(|f| {
+                    if f.name() == SEQ_COLUMN_NAME {
+                        None
+                    } else {
+                        Some(f.clone())
+                    }
+                })
+                .collect_vec();
+            Arc::new(Schema::new_with_metadata(
+                fields,
+                stream.schema().metadata.clone(),
+            ))
+        } else {
+            stream.schema()
+        };
         Self {
             stream,
             num_primary_keys,
             seq_idx,
             value_operator,
+            keep_sequence,
             pending_batch: None,
             arrow_schema,
         }
@@ -312,6 +325,10 @@ impl MergeStream {
 
         let mut output_batches =
             concat_batches(&self.stream.schema(), output_batches.iter()).context("concat batch")?;
+        if self.keep_sequence {
+            return Ok(Some(output_batches));
+        }
+
         // Remove seq column
         output_batches.remove_column(self.seq_idx);
         Ok(Some(output_batches))
@@ -330,7 +347,9 @@ impl Stream for MergeStream {
                 Poll::Pending => return Poll::Pending,
                 Poll::Ready(None) => {
                     let value = if let Some(mut pending) = self.pending_batch.take() {
-                        pending.remove_column(self.seq_idx);
+                        if !self.keep_sequence {
+                            pending.remove_column(self.seq_idx);
+                        }
                         let res = self
                             .value_operator
                             .merge(pending)
@@ -406,6 +425,7 @@ impl ParquetReader {
         ssts: Vec<SstFile>,
         projections: Option<Vec<usize>>,
         predicates: Vec<Expr>,
+        keep_sequence: bool,
     ) -> Result<Arc<dyn ExecutionPlan>> {
         // we won't use url for selecting object_store.
         let dummy_url = ObjectStoreUrl::parse("empty://").unwrap();
@@ -452,6 +472,7 @@ impl ParquetReader {
                     Arc::new(BytesMergeOperator::new(self.schema.value_idxes.clone()))
                 }
             },
+            keep_sequence,
         );
         Ok(Arc::new(merge_exec))
     }
@@ -523,7 +544,9 @@ mod tests {
             .unwrap(),
         ]);
 
-        let stream = MergeStream::new(stream, 1, 2, merge_op);
+        let stream = MergeStream::new(
+            stream, 1, 2, merge_op, false, // keep_sequence
+        );
         check_stream(Box::pin(stream), expected).await;
     }
 
@@ -559,13 +582,14 @@ mod tests {
                     .collect(),
                 None,
                 vec![],
+                false, // keep_sequence
             )
             .unwrap();
         let display_plan =
             datafusion::physical_plan::display::DisplayableExecutionPlan::new(plan.as_ref())
                 .indent(true);
         assert_eq!(
-            r#"MergeExec: [primary_keys: 1, seq_idx: 2]
+            r#"MergeExec: [primary_keys: 1, seq_idx: 2, keep_sequence: false]
   SortPreservingMergeExec: [pk1@0 ASC, __seq__@2 ASC]
     ParquetExec: file_groups={3 groups: [[mock/data/100.sst], [mock/data/101.sst], [mock/data/102.sst]]}, projection=[pk1, value, __seq__], output_orderings=[[pk1@0 ASC, __seq__@2 ASC], [pk1@0 ASC, __seq__@2 ASC], [pk1@0 ASC, __seq__@2 ASC]]
 "#,
